refactor(planner): split oversized planner files under 30 KB limit
5 oversized files split into focused sub-modules using directory packages (charging/, candidates/, milp/). All 31 planner files now under 30 KB. No code logic changed — pure file extraction and import updates. All 2310 tests pass.

Also: remove stale issue-tracking sections from .github/memories.md.
