+++ custom_components/hsem/planner/charging/__init__.py
@@ -1,0 +1,26 @@
+"""Charge scheduling sub-package.
+
+Re-exports all public charge-scheduling functions so callers that were
+previously importing from ``charge_scheduler`` can import from here
+instead.
+"""
+
+from __future__ import annotations
+
+from custom_components.hsem.planner.charging.arbitrage_charge import (
+    apply_arbitrage_grid_charge,
+)
+from custom_components.hsem.planner.charging.opportunistic_charge import (
+    apply_opportunistic_charge,
+)
+from custom_components.hsem.planner.charging.pre_charge import (
+    _apply_grid_charge,
+    apply_charge_schedules,
+)
+
+__all__ = [
+    "apply_arbitrage_grid_charge",
+    "apply_charge_schedules",
+    "apply_opportunistic_charge",
+    "_apply_grid_charge",
+]
