+++ custom_components/hsem/planner/engine_core.py
@@ -8,27 +8,26 @@ directly testable with plain ``pytest`` without a running HA instance.
 
 from __future__ import annotations
 
-from datetime import datetime, timedelta
+from datetime import datetime
 
-from custom_components.hsem.models.data_quality import DataQuality
 from custom_components.hsem.models.ev_config import EVConfig
 from custom_components.hsem.models.planner_input import PlannerInput
 from custom_components.hsem.models.planner_output import PlannerOutput
-from custom_components.hsem.models.time_series import TimeSeriesIndex
 from custom_components.hsem.planner.candidate_generator import (
     CANDIDATE_MILP,
     generate_candidates,
 )
 from custom_components.hsem.planner.candidate_selector import (
-    ev_future_charge_value_per_kwh,
     replacement_price_from_next_discharge,
     select_best_candidate,
 )
-from custom_components.hsem.planner.charge_scheduler import (
+from custom_components.hsem.planner.charging.arbitrage_charge import (
     apply_arbitrage_grid_charge,
-    apply_charge_schedules,
+)
+from custom_components.hsem.planner.charging.opportunistic_charge import (
     apply_opportunistic_charge,
 )
+from custom_components.hsem.planner.charging.pre_charge import apply_charge_schedules
 from custom_components.hsem.planner.cost_function import CostWeights, score_plan
 from custom_components.hsem.planner.discharge_scheduler import (
     apply_discharge_schedules,
@@ -36,15 +35,24 @@ from custom_components.hsem.planner.discharge_scheduler import (
     apply_optimization_strategy,
     calculate_required_battery_until_solar,
 )
+from custom_components.hsem.planner.engine_ev import (
+    _build_and_inject_for_ev,
+    _compute_ev_charger_power,
+)
+from custom_components.hsem.planner.engine_ev_milp import (
+    _build_ev_configs_for_milp,
+)
 from custom_components.hsem.planner.engine_explanation import (
     _build_explanation,
     _derive_windows,
 )
+from custom_components.hsem.planner.engine_population import (
+    _inject_live_data_into_current_slot,
+    _parse_now,
+    _populate_slots,
+)
 from custom_components.hsem.planner.ev_planner import (
     EVChargingPlan,
-    EVPlannerInput,
-    apply_ev_planned_load_to_slots,
-    build_ev_charging_plan,
     rebuild_ev_plan_from_slots,
 )
 from custom_components.hsem.planner.slot_population import (
@@ -52,11 +60,8 @@ from custom_components.hsem.planner.slot_population import (
     build_time_series_index,
     mark_time_passed,
     populate_battery_capacity,
-    populate_consumption,
     populate_estimated_cost,
     populate_net_consumption,
-    populate_prices,
-    populate_solcast,
     usable_capacity,
 )
 from custom_components.hsem.utils.datetime_utils import as_tz
@@ -66,234 +71,6 @@ from custom_components.hsem.utils.misc import (
     clamp_efficiency,
 )
 from custom_components.hsem.utils.recommendations import Recommendations
-
-
-def _parse_now(now_iso: str) -> datetime:
-    """Parse a timezone-aware ISO-8601 string."""
-    dt = datetime.fromisoformat(now_iso)
-    if dt.tzinfo is None:
-        raise ValueError(f"now_iso must be timezone-aware, got: {now_iso!r}")
-    return dt
-
-
-def _populate_slots(
-    slots: list,
-    inp: PlannerInput,
-    tsi: TimeSeriesIndex,
-    warnings: list[str],
-    missing_inputs: list[str],
-) -> tuple[DataQuality, list[str], list[str]]:
-    """Populate price/PV/consumption data, data-quality diagnostics, confidence decay."""
-    log_planner(
-        "debug",
-        "[core] _populate_slots  slots=%d  interval=%dmin  horizon=%dh",
-        len(slots),
-        inp.interval_minutes,
-        inp.interval_length_hours,
-    )
-    populate_prices(slots, inp.price_points, tsi)
-    populate_solcast(
-        slots,
-        inp.solcast_slots,
-        inp.interval_minutes,
-        tsi,
-        corrector=inp.solar_corrector,
-    )
-    populate_consumption(
-        slots,
-        inp.consumption_averages,
-        inp.weight_1d,
-        inp.weight_3d,
-        inp.weight_7d,
-        inp.weight_14d,
-        inp.interval_minutes,
-        tsi,
-    )
-    for hour in sorted(tsi.missing_hours()):
-        missing_inputs.append(f"hour_{hour:02d}")
-    horizon_has_tomorrow = tsi.has_tomorrow_slots()
-    horizon_days = tsi.horizon_days
-    dq = DataQuality(
-        tomorrow_price_missing_hours=sorted(tsi.missing_future_day_price_hours(1)),
-        tomorrow_pv_missing_hours=sorted(tsi.missing_future_day_pv_hours(1)),
-        day2_price_missing_hours=sorted(tsi.missing_future_day_price_hours(2)),
-        day2_pv_missing_hours=sorted(tsi.missing_future_day_pv_hours(2)),
-        today_price_missing_hours=sorted(
-            {
-                m.hour
-                for k in tsi.missing_price_slots
-                if k.day_offset == 0
-                for m in tsi.slots
-                if m.key == k
-            }
-        ),
-        today_pv_missing_hours=sorted(
-            {
-                m.hour
-                for k in tsi.missing_pv_slots
-                if k.day_offset == 0
-                for m in tsi.slots
-                if m.key == k
-            }
-        ),
-        horizon_has_tomorrow=horizon_has_tomorrow,
-        horizon_days=horizon_days,
-    )
-    for tm_label, tm_list in [
-        ("tomorrow_price_missing_hours", dq.tomorrow_price_missing_hours),
-        ("tomorrow_pv_missing_hours", dq.tomorrow_pv_missing_hours),
-        ("day2_price_missing_hours", dq.day2_price_missing_hours),
-        ("day2_pv_missing_hours", dq.day2_pv_missing_hours),
-    ]:
-        if tm_list:
-            hs = ",".join(f"{h:02d}" for h in tm_list)
-            missing_inputs.append(f"{tm_label}:{hs}")
-            labels = {
-                "tomorrow_price_missing_hours": "price",
-                "tomorrow_pv_missing_hours": "PV",
-                "day2_price_missing_hours": "price",
-                "day2_pv_missing_hours": "PV",
-            }
-            warnings.append(
-                f"{'Day+2' if tm_label.startswith('day2') else 'Tomorrow'} {labels[tm_label]} data missing for {len(tm_list)} hour(s): {hs}."
-            )
-    _DECAY_BY_DAY: dict[int, float] = {0: 1.00, 1: 0.90, 2: 0.80}
-    if horizon_days > 1:
-        for slot in slots:
-            si = tsi.slot_index_for(slot.start)
-            if si is None:
-                continue
-            do = tsi.slots[si].key.day_offset
-            if do > 0:
-                slot.solcast_pv_estimate_kwh = round(
-                    slot.solcast_pv_estimate_kwh * _DECAY_BY_DAY.get(do, 0.80), 3
-                )
-        if horizon_days >= 2:
-            warnings.append(
-                f"Multi-day horizon ({horizon_days} day(s)): confidence decay applied to PV estimates."
-            )
-    log_planner(
-        "debug",
-        "[core] _populate_slots DONE  warnings=%d  missing=%d  horizon_days=%d  "
-        "has_tomorrow=%s",
-        len(warnings),
-        len(missing_inputs),
-        horizon_days,
-        horizon_has_tomorrow,
-    )
-
-    return dq, warnings, missing_inputs
-
-
-def _inject_live_data_into_current_slot(
-    slots: list,
-    inp: PlannerInput,
-    now: datetime,
-) -> None:
-    """Replace forecast PV and consumption in the current slot with live measurements.
-
-    The current (partially-elapsed) slot's ``solcast_pv_estimate_kwh`` and
-    ``avg_house_consumption_kwh`` are overwritten with values derived from
-    live power sensors.  This ensures the MILP and all candidates use
-    measured (not forecast) data for the slot that is already in progress.
-
-    The live power in Watts is converted to kWh by multiplying by the
-    slot's full duration in hours.  This gives the *projected* full-slot
-    energy if the current power level were sustained for the entire slot.
-    The MILP's energy balance constraint then correctly accounts for the
-    remaining portion of the slot.
-
-    Args:
-        slots: Fully populated slot list (after ``_populate_slots``).
-        inp: Planner input containing live power readings.
-        now: Timezone-aware current datetime.
-    """
-    slot_hours = inp.interval_minutes / 60.0
-
-    for slot in slots:
-        s_start = as_tz(slot.start, now.tzinfo)
-        s_end = as_tz(slot.end, now.tzinfo)
-        if s_start <= now < s_end:
-            # Convert live Watts to projected full-slot kWh.
-            if inp.live_solar_production_w > 1e-9:
-                live_pv_kwh = (inp.live_solar_production_w / 1000.0) * slot_hours
-                log_planner(
-                    "debug",
-                    "[core] _inject_live_data  slot=%s  "
-                    "pv: forecast=%.3f → live=%.3f kWh",
-                    slot.start.isoformat(),
-                    slot.solcast_pv_estimate_kwh,
-                    live_pv_kwh,
-                )
-                slot.solcast_pv_estimate_kwh = round(live_pv_kwh, 3)
-
-            if inp.live_house_consumption_w > 1e-9:
-                # When house power includes EV charger load, the live reading
-                # may include EV power that the battery must not serve.
-                #
-                # Layer 1: if ev_session_charge_kw is available (EV actively
-                # charging with a known power reading), subtract it directly.
-                #
-                # Layer 2: otherwise, if the live reading is dramatically
-                # higher than the forecast (>3×), cap the injection at the
-                # forecast value.  A 3× spike is unambiguous EV load — normal
-                # house load does not triple between slots (issue #592).
-                live_house_w = inp.live_house_consumption_w
-                if inp.house_power_includes_ev:
-                    ev_ac_w = 0.0
-                    if (
-                        inp.ev_session_charge_kw is not None
-                        and inp.ev_session_charge_kw > 1e-9
-                    ):
-                        ev_ac_w += inp.ev_session_charge_kw * 1000.0
-                    if (
-                        inp.ev_second_session_charge_kw is not None
-                        and inp.ev_second_session_charge_kw > 1e-9
-                    ):
-                        ev_ac_w += inp.ev_second_session_charge_kw * 1000.0
-                    live_house_w = max(live_house_w - ev_ac_w, 0.0)
-
-                live_load_kwh = (live_house_w / 1000.0) * slot_hours
-
-                # If the live reading (after subtracting known EV power) still
-                # exceeds the forecast by >3×, it likely contains unmeasured EV
-                # load.  Cap at the forecast to prevent the battery from
-                # serving unknown EV demand.
-                forecast = slot.avg_house_consumption_kwh
-                if (
-                    inp.house_power_includes_ev
-                    and live_load_kwh > forecast * 3.0
-                    and forecast > 1e-9
-                ):
-                    log_planner(
-                        "debug",
-                        "[core] _inject_live_data  slot=%s  "
-                        "live %.3f kWh > 3× forecast %.3f kWh — "
-                        "capping at forecast (probable unmeasured EV load)",
-                        slot.start.isoformat(),
-                        live_load_kwh,
-                        forecast,
-                    )
-                    live_load_kwh = forecast
-
-                log_planner(
-                    "debug",
-                    "[core] _inject_live_data  slot=%s  "
-                    "load: forecast=%.3f → live=%.3f kWh",
-                    slot.start.isoformat(),
-                    slot.avg_house_consumption_kwh,
-                    live_load_kwh,
-                )
-                slot.avg_house_consumption_kwh = round(live_load_kwh, 3)
-
-            # Also update the sub-window averages so they stay consistent
-            # with the main average (used by spike detection in
-            # populate_consumption).
-            slot.avg_house_consumption_1d_kwh = slot.avg_house_consumption_kwh
-            slot.avg_house_consumption_3d_kwh = slot.avg_house_consumption_kwh
-            slot.avg_house_consumption_7d_kwh = slot.avg_house_consumption_kwh
-            slot.avg_house_consumption_14d_kwh = slot.avg_house_consumption_kwh
-            break
 
 
 def _schedule_slots(
@@ -405,189 +182,6 @@ def _schedule_slots(
     return mcps, mdps, max_soc_kwh, rc, warnings
 
 
-def _compute_ev_charger_power(
-    slots: list,
-    slot_starts: list[datetime],
-    ev_plan: EVChargingPlan | None,
-    interval_minutes: int,
-    now: datetime,
-    *,
-    second: bool = False,
-) -> None:
-    """Compute per-slot EV charger target power (W) and write to slots.
-
-    ``EVChargingSlot.ac_load_kwh`` is the AC-side energy the charger draws
-    from the grid/PV.  The target power is::
-
-        AC power (W) = (ac_load_kwh / slot_hours) × 1000
-
-    For the **current** (partially elapsed) slot the divisor is the
-    remaining slot duration, not the full slot width, because the EV
-    planner already scales ``ac_load_kwh`` to the remaining minutes.
-    Using the full slot width would understate the required charge power.
-
-    When the plan is ``None`` or empty the field stays at the default 0.0.
-
-    Args:
-        slots: Mutable planner slot list to update in place.
-        slot_starts: Slot start datetimes (same length as *slots*).
-        ev_plan: EV charging plan (may be ``None``).
-        interval_minutes: Slot width in minutes.
-        now: Current time (timezone-aware), used to detect the current slot.
-        second: If ``True``, write to ``ev_second_charger_calculated_power``;
-            otherwise write to ``ev_charger_calculated_power``.
-    """
-    if ev_plan is None or not ev_plan.charging_slots:
-        return
-
-    # Build a lookup from UTC key → slot index.
-    from custom_components.hsem.utils.datetime_utils import utc_key
-
-    slot_map = {utc_key(s): i for i, s in enumerate(slot_starts)}
-    full_hours = interval_minutes / 60.0
-
-    for ev_slot in ev_plan.charging_slots:
-        idx = slot_map.get(utc_key(ev_slot.start))
-        if idx is None:
-            continue
-        if ev_slot.ac_load_kwh < 1e-9:
-            continue
-
-        # For the current (partially elapsed) slot, the EV planner has
-        # already scaled ``ac_load_kwh`` to the *remaining* minutes.
-        # Divide by remaining hours to get the correct target power.
-        # For future slots the full slot width is used.
-        slot_end = slots[idx].end
-        if slots[idx].start <= now < slot_end:
-            remaining_min = max((slot_end - now).total_seconds() / 60.0, 0.0167)
-            slot_hours = remaining_min / 60.0
-        else:
-            slot_hours = full_hours
-
-        ac_power_w = round((ev_slot.ac_load_kwh / slot_hours) * 1000)
-
-        # Cap at the charger's rated AC power — the EV planner may allocate
-        # a full slot's worth of energy to a slot with only a few minutes
-        # remaining.  The charger physically cannot exceed its nameplate.
-        if ev_plan.charger_power_kw > 1e-9:
-            max_ac_power_w = round(ev_plan.charger_power_kw * 1000)
-            ac_power_w = min(ac_power_w, max_ac_power_w)
-
-        # Floor at the charger's minimum operating power — if the target
-        # power is below the minimum the charger needs to start, it will
-        # never deliver any energy.  Zero out the field so the applier
-        # does not attempt to throttle the charger below its minimum.
-        if (
-            ev_plan.charger_min_power_w > 1e-9
-            and ac_power_w < ev_plan.charger_min_power_w
-        ):
-            ac_power_w = 0
-
-        attr = (
-            "ev_second_charger_calculated_power"
-            if second
-            else "ev_charger_calculated_power"
-        )
-        setattr(slots[idx], attr, ac_power_w)
-
-
-def _build_and_inject_for_ev(
-    enabled: bool,
-    connected: bool,
-    smart: bool,
-    soc: float,
-    target: float,
-    cap_kwh: float,
-    pwr_kw: float,
-    eff: float,
-    min_pwr_w: float,
-    deadline: datetime | None,
-    base_includes: bool,
-    allow_past_target: bool,
-    label: str,
-    now: datetime,
-    slots: list,
-    slot_starts: list,
-    slot_ends: list,
-    slot_prices: list,
-    slot_net_surplus: list[float],
-    combined_ev_raw_load: list[float],
-    combined_ev_injected_load: list[float],
-    warnings: list[str],
-) -> EVChargingPlan | None:
-    """Build an EV charging plan and accumulate its loads."""
-    if not enabled:
-        return None
-    log_planner(
-        "debug",
-        "[core] _build_and_inject_for_ev  label=%s  connected=%s  smart=%s  "
-        "soc=%.1f%%  target=%.1f%%  cap=%.2f  pwr=%.2f  eff=%.1f%%  min_pwr=%.0fW",
-        label,
-        connected,
-        smart,
-        soc,
-        target,
-        cap_kwh,
-        pwr_kw,
-        eff,
-        min_pwr_w,
-    )
-    ev_inp = EVPlannerInput(
-        enabled=enabled,
-        ev_connected=connected,
-        smart_charging_enabled=smart,
-        current_soc_pct=soc,
-        target_soc_pct=target,
-        battery_capacity_kwh=cap_kwh,
-        charger_power_kw=pwr_kw,
-        charger_efficiency_pct=eff,
-        charger_min_power_w=min_pwr_w,
-        deadline=deadline,
-        base_load_includes_ev=base_includes,
-        allow_charge_past_target_soc=allow_past_target,
-        now=now,
-    )
-    plan = build_ev_charging_plan(
-        ev_inp,
-        slots_start=slot_starts,
-        slots_end=slot_ends,
-        slot_net_surplus_kwh=slot_net_surplus,
-        slot_import_price=slot_prices,
-    )
-    raw = [0.0] * len(slots)
-    apply_ev_planned_load_to_slots(
-        slot_starts=slot_starts,
-        slot_ev_planned_load_kwh=raw,
-        ev_plan=plan,
-        base_load_includes_ev=False,
-    )
-    for i in range(len(slots)):
-        combined_ev_raw_load[i] += raw[i]
-    inj = [0.0] * len(slots)
-    apply_ev_planned_load_to_slots(
-        slot_starts=slot_starts,
-        slot_ev_planned_load_kwh=inj,
-        ev_plan=plan,
-        base_load_includes_ev=base_includes,
-    )
-    for i in range(len(slots)):
-        combined_ev_injected_load[i] += inj[i]
-    if plan.state not in ("not_connected", "smart_charging_disabled", "fully_charged"):
-        warnings.append(
-            f"EV planned load ({label}): state={plan.state}, total_kwh_needed={plan.total_kwh_needed:.2f}, charging_slots={len(plan.charging_slots)}, base_load_includes_ev={base_includes}."
-        )
-    log_planner(
-        "debug",
-        "[core] _build_and_inject_for_ev DONE  label=%s  state=%s  slots=%d  "
-        "total_kwh=%.3f",
-        label,
-        plan.state,
-        len(plan.charging_slots),
-        plan.total_kwh_needed,
-    )
-    return plan
-
-
 def _select_candidate(
     slots: list,
     inp: PlannerInput,
@@ -650,188 +244,6 @@ def _select_candidate(
     return candidates, winner, rejected, hyst
 
 
-def _build_ev_configs_for_milp(
-    inp: PlannerInput,
-    slots: list,
-    now: datetime,
-) -> list[EVConfig] | None:
-    """Build EVConfig list for the MILP from PlannerInput EV fields.
-
-    Maps the user-configured deadline (clamped by the one-midnight-crossing
-    horizon cap) to an LP slot index.  Returns ``None`` when no EVs are
-    active or neither EV has sufficient config to be optimised.
-    """
-
-    def _effective_deadline_dt(user_deadline: datetime | None) -> datetime:
-        """Replicate ev_planner._effective_deadline without HA import."""
-        horizon_cap = (now + timedelta(days=2)).replace(
-            hour=0, minute=0, second=0, microsecond=0
-        )
-        if user_deadline is None:
-            return horizon_cap
-        return min(user_deadline, horizon_cap)
-
-    configs: list[EVConfig] = []
-    future_slots = [i for i, s in enumerate(slots) if as_tz(s.end, now.tzinfo) > now]
-    if not future_slots:
-        return None
-
-    # Avoided-future-import-cost valuation for charge-past-target EVs
-    # (issue #630). Computed once — depends only on the slot price
-    # forecast, not per-EV state. Per-EV confidence factors are applied
-    # below when building each EVConfig.
-    ev_avg_future_import_price = ev_future_charge_value_per_kwh(
-        slots, now, confidence_factor=1.0
-    )
-
-    # Build config for each EV slot pair (primary, secondary).
-    # Each source tuple carries an is_second flag so the downstream
-    # write-out loop can route EV power to the correct output field by
-    # identity rather than by list position (issue #646).
-    ev_sources: list[
-        tuple[
-            bool,
-            bool,
-            bool,
-            float,
-            float,
-            float,
-            float,
-            float,
-            float,
-            datetime | None,
-            bool,
-            bool,
-            float,
-            bool,
-        ]
-    ] = [
-        (
-            inp.ev_planned_load_enabled,
-            inp.ev_planned_load_connected,
-            inp.ev_planned_load_smart_charging_enabled,
-            inp.ev_planned_load_current_soc_pct,
-            inp.ev_planned_load_target_soc_pct,
-            inp.ev_planned_load_battery_capacity_kwh,
-            inp.ev_planned_load_charger_power_kw,
-            inp.ev_planned_load_charger_efficiency_pct,
-            inp.ev_planned_load_charger_min_power_w,
-            inp.ev_planned_load_deadline,
-            inp.ev_planned_load_base_load_includes_ev,
-            inp.ev_planned_allow_charge_past_target_soc,
-            inp.ev_past_target_confidence_factor,
-            False,  # is_second = False (primary EV)
-        ),
-        (
-            inp.ev_second_planned_load_enabled,
-            inp.ev_second_planned_load_connected,
-            inp.ev_second_planned_load_smart_charging_enabled,
-            inp.ev_second_planned_load_current_soc_pct,
-            inp.ev_second_planned_load_target_soc_pct,
-            inp.ev_second_planned_load_battery_capacity_kwh,
-            inp.ev_second_planned_load_charger_power_kw,
-            inp.ev_second_planned_load_charger_efficiency_pct,
-            inp.ev_second_planned_load_charger_min_power_w,
-            inp.ev_second_planned_load_deadline,
-            inp.ev_second_planned_load_base_load_includes_ev,
-            inp.ev_second_allow_charge_past_target_soc,
-            inp.ev_second_past_target_confidence_factor,
-            True,  # is_second = True (second EV)
-        ),
-    ]
-    for (
-        enabled,
-        connected,
-        smart,
-        soc_pct,
-        target_pct,
-        cap,
-        pwr,
-        eff_pct,
-        min_pwr_w,
-        deadline,
-        base_includes,
-        allow_past_target,
-        past_target_confidence_factor,
-        is_second,
-    ) in ev_sources:
-        if not enabled:
-            continue
-        if not connected or not smart:
-            continue
-        if cap <= 1e-9 or pwr <= 1e-9:
-            continue
-        initial_kwh = (soc_pct / 100.0) * cap
-        target_kwh = (target_pct / 100.0) * cap
-
-        # When the EV is already at or above its target SoC, normally we
-        # skip it — there is no energy deficit to meet.  But when
-        # allow_charge_past_target_soc is enabled and the EV is not yet
-        # at 100 %, the MILP should still include the EV so it can
-        # allocate surplus PV that would otherwise be curtailed or
-        # exported at low/negative prices.  In this mode the deadline
-        # constraint is suppressed (deadline_slot=None) so the MILP
-        # never imports from grid to meet a target that is already
-        # satisfied — it only charges from free/cheap surplus.
-        at_or_above_target = target_kwh <= initial_kwh + 1e-9
-        deadline_slot: int | None = None
-        charge_past_target = False
-        if at_or_above_target:
-            if not allow_past_target or soc_pct >= 100:
-                continue  # fully charged or past-target not allowed
-            # Charge-past-target mode: allow up to 100 %, no deadline pressure.
-            target_kwh = cap
-            charge_past_target = True
-        else:
-            # Normal mode: map deadline to LP slot index.
-            eff_deadline = _effective_deadline_dt(deadline)
-            for lp_t, slot_i in enumerate(future_slots):
-                s = slots[slot_i]
-                if as_tz(s.end, now.tzinfo) <= eff_deadline:
-                    deadline_slot = lp_t
-                else:
-                    break
-            if deadline_slot is None:
-                # No slot before deadline
-                continue
-            charge_past_target = False
-
-        eff = max(eff_pct, 1.0) / 100.0
-        slot_hours = inp.interval_minutes / 60.0
-        max_dc = pwr * slot_hours * eff  # DC-side kWh per slot
-
-        future_value_per_kwh: float | None = None
-        if charge_past_target and ev_avg_future_import_price is not None:
-            future_value_per_kwh = (
-                past_target_confidence_factor * ev_avg_future_import_price
-            )
-
-        configs.append(
-            EVConfig(
-                enabled=True,
-                initial_soc_kwh=round(initial_kwh, 3),
-                target_kwh=round(target_kwh, 3),
-                capacity_kwh=round(cap, 3),
-                max_charge_per_slot=round(max_dc, 4),
-                charger_efficiency=round(eff, 4),
-                charger_min_power_w=round(min_pwr_w, 1),
-                deadline_slot=deadline_slot,
-                base_load_includes_ev=base_includes,
-                charge_past_target=charge_past_target,
-                future_value_per_kwh=future_value_per_kwh,
-                is_second=is_second,
-            )
-        )
-
-    # Apply session charge power from live EV state (issue #615).
-    # Route by identity (is_second), not list position (issue #646).
-    for cfg in configs:
-        if cfg.is_second:
-            cfg.session_charge_kw = inp.ev_second_session_charge_kw
-        else:
-            cfg.session_charge_kw = inp.ev_session_charge_kw
-
-    return configs if configs else None
 
 
 def run_planner(inp: PlannerInput) -> PlannerOutput:
