+++ custom_components/hsem/planner/milp/__init__.py
@@ -1,0 +1,9 @@
+"""MILP optimizer sub-package."""
+
+from custom_components.hsem.planner.milp_optimizer import (
+    CANDIDATE_MILP,
+    is_scipy_available,
+    solve_milp,
+)
+
+__all__ = ["solve_milp", "CANDIDATE_MILP", "is_scipy_available"]
